+++ javascript/node/selenium-webdriver/example/chrome_headless.js
@@ -1,0 +1,47 @@
+// Licensed to the Software Freedom Conservancy (SFC) under one
+// or more contributor license agreements.  See the NOTICE file
+// distributed with this work for additional information
+// regarding copyright ownership.  The SFC licenses this file
+// to you under the Apache License, Version 2.0 (the
+// "License"); you may not use this file except in compliance
+// with the License.  You may obtain a copy of the License at
+//
+//   http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing,
+// software distributed under the License is distributed on an
+// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+// KIND, either express or implied.  See the License for the
+// specific language governing permissions and limitations
+// under the License.
+
+/**
+ * @fileoverview An example of running Chrome in headless mode.
+ *
+ * Before running this script, ensure you have Chrome 59+ installed and that
+ * chromedriver 2.30+ is present on your system PATH:
+ * <https://sites.google.com/a/chromium.org/chromedriver/downloads>
+ *
+ * Usage:
+ *   node selenium-webdriver/example/chrome_headless.js
+ */
+
+const chrome = require('../chrome');
+const {Builder, By, until} = require('..');
+
+const width = 640;
+const height = 480;
+
+let driver = new Builder()
+    .forBrowser('chrome')
+    .setChromeOptions(
+        new chrome.Options().headless().windowSize({width, height}))
+    .build();
+
+driver.get('http://www.google.com/ncr')
+    .then(_ => driver.findElement(By.name('q')).sendKeys('webdriver'))
+    .then(_ => driver.findElement(By.name('btnG')).click())
+    .then(_ => driver.wait(until.titleIs('webdriver - Google Search'), 1000))
+    .then(
+        _ => driver.quit(),
+        e => driver.quit().then(() => { throw e; }));